Fix compiler warnings (#17)
* Enable only basic warnings

* Fix basic warnings

* Enable deprecated-copy-dtor

* Fix deprecated-copy-dtor

* Enable -Wsign-conversion

* Fix -Wsign-conversion

* Enable -Wconversion

* Fix -Wconversion

* Enable -Wnull-dereference

* Fix -Wnull-dereference

* Enable -Wsuggest-attribute=const

* Fix -Wsuggest-attribute=const

* Enable -Wsuggest-attribute=pure

* Fix -Wsuggest-attribute=pure

* Enable -Weffc++

* WIP: Fix -Weffc++

* Fix -Weffc++

* Fix formatting after changes

+++ apps/ogre/libs/libcli/test/cli/Command.cpp
@@ -10,12 +10,11 @@
 
 TEST_CASE("tasklib/cli/Command") {
   // Setup
-  std::srand(std::time(nullptr));
+  std::srand(static_cast<unsigned>(std::time(nullptr)));
   int result = rand();
   //  Command
   ogre::command cmd("git");
-  cmd.add_action(
-      [result](ogre::parameters const& /*args*/) -> int { return result; });
+  cmd.add_action([result](ogre::parameters const& /*args*/) -> int { return result; });
   //  Input
   std::vector<std::string_view> input{"git"};
 
@@ -47,8 +46,7 @@ TEST_CASE("tasklib/cli/Command option") {
   ogre::command cmd("git");
   auto& rm = cmd.add_subcommand("rm");
   rm.add_option("-f", "--force").add_help("Override the up-to-date check.");
-  rm.add_option("-r").add_help(
-      "Allow recursive removal when leading directory name is given.");
+  rm.add_option("-r").add_help("Allow recursive removal when leading directory name is given.");
   rm.add_action([result](ogre::parameters const& p_args) -> int {
     REQUIRE(p_args.options.contains("-f"));
     CHECK(p_args.options.at("-f") == "true");
